% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_cluster_utils.R
\name{annotate_gene_clusters}
\alias{annotate_gene_clusters}
\title{Annotates genes clusters using g:OSt}
\usage{
annotate_gene_clusters(
  moduli,
  organism,
  n.terms = 3,
  user_threshold = 0.01,
  ...
)
}
\arguments{
\item{moduli}{A moduli object}

\item{organism}{Organism name, constructed by concatenating the fist letter of the
name with the family name}

\item{user_threshold}{p-value threshold for significance}

\item{...}{Additional arguments passed to \link[gprofiler2]{gost}}

\item{n.tems}{Number of terms to save}
}
\value{
A moduli object with the significant term names and associated p-values
saved to moduli$gene.clusters$term.names and moduli$gene.clusters$term.p.values
respectively, ordered by significance.
}
\description{
Enriches gene clusters with most significant terms according to g:OSt by
using \link[gprofiler2]{gost}.
}
