% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_clustering.R
\name{cluster_group_metadata}
\alias{cluster_group_metadata}
\title{Retrieves metadata associated to analysis cluster groups}
\usage{
cluster_group_metadata(moduli, thld = 0.05)
}
\arguments{
\item{moduli}{A moduli object with a dendrogram representation}

\item{thld}{Significance threshold}
}
\value{
A data.frame with ids of analysis cluster groups, their heights, an
indicator if they are leafs (i.e. correspond to a single analysis cluster), number
of points in the moduli present in the analysis cluster group, differentially 
expressed and suppressed gene clusters with associated p-values (Fisher's exact test).
Cluster groups are ordered by depth first search.
}
\description{
Retrieves metadata associated to analysis cluster groups
}
\examples{
data("pbmc_small_moduli")

# clustering moduli space
pbmc_small_moduli <- get_snn(pbmc_small_moduli, 4)
pbmc_small_moduli <- cluster_moduli_space(pbmc_small_moduli)

pbmc_small_moduli <- organize_analysis_clusters(pbmc_small_moduli)
cluster_group_metadata(pbmc_small_moduli)

}
