% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduli_clustering.R
\name{enrich_analysis_clusters}
\alias{enrich_analysis_clusters}
\title{Enriches analysis clusters with differentially expressed gene clusters}
\usage{
enrich_analysis_clusters(moduli, thld = 0.05)
}
\arguments{
\item{moduli}{A moduli object with analysis clusters save to the \code{analysis.clusters} slot}

\item{thld}{Significance threshold}
}
\value{
A moduli object with diffentially expressed gene clusters and associated p-values
saved in the \code{analysis.clusters$exp.gene.clusters} and \code{analysis.clusters$enrichment.p.vals}
respectively, ordered by of significance.
}
\description{
Applies Fisher's exact test to find differentially expressed gene clusters in
each analysis cluster.
}
