% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_cluster_utils.R
\name{gene_cluster_metadata}
\alias{gene_cluster_metadata}
\title{Retrieves gene cluster metadata}
\usage{
gene_cluster_metadata(moduli)
}
\arguments{
\item{moduli}{A moduli object}
}
\value{
A data.frame with the ids of gene clusters, their sizes, names of expressed
terms and associated p-values (if present), laplacian score and associated rank (if present)
and ids of analysis clusters where they are differentially expressed (if present).
}
\description{
Retrieves gene cluster metadata
}
\examples{
data("pbmc_small_moduli")
gene_cluster_metadata(pbmc_small_moduli)

}
