% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_clustering.R
\name{get_analysis_cluster_groups}
\alias{get_analysis_cluster_groups}
\title{Retrieves analysis cluster groups}
\usage{
get_analysis_cluster_groups(moduli)
}
\arguments{
\item{moduli}{A moduli object with a dendrogram representation}
}
\value{
A data.frame with ids of analysis cluster groups, their heights, an
indicator if they are leafs (i.e. correspond to a single analysis cluster) and
analysis clusters associated to each group. Cluster groups are ordered
by depth first search.
}
\description{
Retrieves analysis cluster groups
}
\examples{
data("pbmc_small_moduli")

# clustering moduli space
pbmc_small_moduli <- get_snn(pbmc_small_moduli, 4)
pbmc_small_moduli <- cluster_moduli_space(pbmc_small_moduli)

pbmc_small_moduli <- organize_analysis_clusters(pbmc_small_moduli)
get_analysis_cluster_groups(pbmc_small_moduli)

}
