% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_medioids}
\alias{get_medioids}
\title{Retrieves medioids of sets of points}
\usage{
get_medioids(moduli, point.sets = NULL, analysis.clusters = NULL)
}
\arguments{
\item{moduli}{A moduli object}

\item{point.sets}{Sets of points as a list of integer vectors with point ids}

\item{analysis.clusters}{A collection of analysis clusters as an integer vector with ids}
}
\value{
A vector with ids of medioids
}
\description{
Finds the medioids of a collection of sets of points or of a collection of 
analysis clusters.
}
\examples{
data("pbmc_small_moduli")

# clustering is not necessary, but it shows an application
pbmc_small_moduli <- get_snn(pbmc_small_moduli, 4)
pbmc_small_moduli <- cluster_moduli_space(pbmc_small_moduli)

# retrieving medioids of all analysis clusters
a.clst.ids <- analysis_cluster_metadata(pbmc_small_moduli)$id
medioids <- get_medioids(pbmc_small_moduli, analysis.clusters = a.clst.ids)


}
