% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_snn}
\alias{get_snn}
\title{Enriches moduli with its shared nearest neighbors graph}
\usage{
get_snn(moduli, k, thld = 0)
}
\arguments{
\item{moduli}{A moduli object}

\item{k}{Number of nearest neighbors}

\item{thld}{Minimum of edges weights to be kept}
}
\value{
A moduli object with an snn graph in the \code{snn.graph} slot.
}
\description{
Enriches moduli with its shared nearest neighbors graph
}
