\name{moduli-package}
\alias{moduli-package}
\alias{moduli}
\docType{package}
\title{
\packageTitle{moduli}
}
\description{
\packageDescription{moduli}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{moduli}
\packageIndices{moduli}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{moduli}

Maintainer: \packageMaintainer{moduli}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
