% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_clustering.R
\name{organize_analysis_clusters}
\alias{organize_analysis_clusters}
\title{Organizes the analysis clusters}
\usage{
organize_analysis_clusters(moduli, method = "average", thld = 0.05)
}
\arguments{
\item{moduli}{A moduli object with analysis clusters}

\item{method}{Agglomeration method used (see \link[stats]{hclust}). The default value
is \code{"averege"}.}

\item{thld}{Significance threshold for gene cluster enrichment}
}
\value{
A moduli object with a dendrogram representation of the clusters in the
\code{dendro} slot and associated metadata in the \code{cluster.group.metadata}
slot (see \link{cluster_group_metadata})
}
\description{
Clusters the moduli space hierarchically, starting by the analysis clusters
obtained by cluster_moduli_space.
}
\examples{
data("pbmc_small_moduli")

# clustering moduli space
pbmc_small_moduli <- get_snn(pbmc_small_moduli, 4)
pbmc_small_moduli <- cluster_moduli_space(pbmc_small_moduli)

pbmc_small_moduli <- organize_analysis_clusters(pbmc_small_moduli)


}
