% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{point_metadata}
\alias{point_metadata}
\title{Retrieve point metadata}
\usage{
point_metadata(moduli, points = moduli$points$id)
}
\arguments{
\item{moduli}{A moduli object}

\item{points}{vector with ids of points to be retrieved, will retrieve all points by default}
}
\value{
A data.frame with point ids, gene clusters represented and, if available,
the analysis cluster the point belongs to.
}
\description{
Retrieves metadata of moduli point
}
\examples{
data("pbmc_small_moduli")
point_metadata(pbmc_small_moduli)

}
