% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{retrieve_consensus}
\alias{retrieve_consensus}
\title{Retries consensus metric}
\usage{
retrieve_consensus(
  moduli,
  point.ids = NULL,
  analysis.cluster.ids = NULL,
  desc.file = NULL,
  n.cores = 1,
  seed = 42,
  filebacked = F,
  persist = F,
  verbose = T
)
}
\arguments{
\item{moduli}{A moduli object}

\item{point.ids}{Ids of moduli points to have their associated metrics averaged, default value is NULL}

\item{analysis.cluster.ids}{Ids of analysis clusters to have associated metrics averaged,
default value is NULL. The user must give either point ids or cluster analysis ids}

\item{desc.file}{Optional bigmatrix descriptor file with moduli embeddings as produced by
\link{get_moduli}. See 'Details'}

\item{n.cores}{Numbers of cores to use, default value is 1}

\item{seed}{Random seed to use, default value is 42}

\item{filebacked}{Use a filebacked bigmatrix to store embeddings, only has an effect if
\code{desc.file} is NULL. Default value is FALSE}

\item{persist}{Keep bigmatrix with embeddings, only has effect if \code{filebacked} is TRUE.
Default value is FALSE}

\item{verbose}{Print time stamps, default value is TRUE}
}
\value{
A "dist" object with avereged cell distances and cell labels
}
\description{
Computes the average distance between cell pairs across a set of embeddings
}
\details{
If a descriptor file is given, pre-calculated embeddings will be used.
In this case the only cells represented in the output are the cells in \code{moduli$cells}.
If no descriptor file is given, embedding will be calculated again, but all cells in the
moduli's Seurat object will be represented in the output.
}
\examples{
data("pbmc_small_moduli")

# retrieving consensus metric of all embeddings containing cluster 3
cons.metric <- retrieve_consensus(pbmc_small_moduli,
                                 which_point(pbmc_small_moduli, 3, filter = "is_subset"))
  

}
