% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{retrieve_point}
\alias{retrieve_point}
\title{Retrieves Seurat object associated to a point in the moduli space}
\usage{
retrieve_point(moduli, point.id)
}
\arguments{
\item{moduli}{A moduli object}

\item{point.id}{Id of point}
}
\value{
A Seurat object with the PCA analysis associated to the point in the
\code{reductions} slot and the average quantile-normalized expression level of
each gene cluster in each cell saved in the slot "gene_cluster_id_expr" of the object's
meta data, where "id" is the id of the associated gene cluster.
}
\description{
Runs PCA with features associated to the given point, also enriches
Seurat object with expression levels of each cluster in each cell
}
\examples{
data("pbmc_small_moduli")

# retrieving point representing clusters 1 and 3
pbmc_small <- retrieve_point(pbmc_small_moduli, which_point(pbmc_small_moduli, c(1, 3)))

}
