% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{run_umap_genes}
\alias{run_umap_genes}
\title{Runs UMAP on gene space}
\usage{
run_umap_genes(moduli, metric = NULL, n_neighbors = 15, ..., seed = 123)
}
\arguments{
\item{moduli}{A moduli object}

\item{metric}{Metric in gene space to use as a dist object, the object must have the names of the genes. If
NULL, the correlation metric based on the \code{"scale.data"} slot of \code{moduli$seurat[[moduli$assay]]}
will be used. Default value is NULL}

\item{n_neighbors}{Size of local neighborhood of umap (see \link[uwot]{umap}), defaults to 15}

\item{...}{Additional parameters passed to \link[uwot]{umap}}

\item{seed}{Random seed, default value is 123}
}
\value{
A moduli object with a umap representation saved to the \code{gene.umap} slot
}
\description{
Computes a 2-d UMAP representation of the gene space using \link[uwot]{umap}.
}
\examples{
data("pbmc_small_moduli")

pbmc_small_moduli <- run_umap_genes(pbmc_small_moduli)

}
