% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{run_umap_moduli}
\alias{run_umap_moduli}
\title{Runs UMAP on moduli space}
\usage{
run_umap_moduli(moduli, n_neighbors = 15, ..., seed = 123)
}
\arguments{
\item{moduli}{A moduli object}

\item{n_neighbors}{Size of local neighborhood of umap (see \link[uwot]{umap}), defaults to 15}

\item{...}{Additional parameters passed to \link[uwot]{umap}}

\item{seed}{Random seed, default value is 123}
}
\value{
A moduli object with a umap representation saved to the \code{moduli.umap} slot
}
\description{
Computes a 2-d UMAP representation of the moduli space using \link[uwot]{umap}.
}
\examples{
data("pbmc_small_moduli")

pbmc_small_moduli <- run_umap_moduli(pbmc_small_moduli)

}
