% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_clustering.R
\name{visualize_analysis_dendrogram}
\alias{visualize_analysis_dendrogram}
\title{Plot a dendrogram of the analysis cluster groups}
\usage{
visualize_analysis_dendrogram(moduli, hang = F)
}
\arguments{
\item{moduli}{A moduli object with a dendrogram representation}

\item{hang}{Whether to hang dendrogram leafs, the default value is FALSE. Setting to
TRUE can help distinguish leafs.}
}
\value{
A plotly visualization. Hover text contains about analysis cluster groups. Each leaf
of the dendrogram represents an analysis cluster and is labeled with the respective id.
}
\description{
Create an interactive visualization of analysis cluster groups.
}
\examples{
data("pbmc_small_moduli")

# clustering moduli space
pbmc_small_moduli <- get_snn(pbmc_small_moduli, 4)
pbmc_small_moduli <- cluster_moduli_space(pbmc_small_moduli)

pbmc_small_moduli <- organize_analysis_clusters(pbmc_small_moduli)
visualize_analysis_dendrogram(pbmc_small_moduli)

}
