% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualize_gene_space}
\alias{visualize_gene_space}
\title{Interactive UMAP plot of gene space}
\usage{
visualize_gene_space(
  moduli,
  color.clusters = NULL,
  mark.genes = NULL,
  ignore.case = F,
  title = "UMAP plot of gene space"
)
}
\arguments{
\item{moduli}{A moduli object with a umap representation of gene space}

\item{color.clusters}{Vector with ids of gene clusters to be colored, can only color 12 clusters. If NULL
the first 12 gene clusters are colored. Default value is NULL}

\item{mark.genes}{Integer vector with names of genes to mark with "x", all points are marked with a circle if NULL.
Default value is NULL}

\item{ignore.case}{Whether to ignore case for genes given to mark.genes, default value is FALSE}

\item{title}{Title of plot, default value is "UMAP plot of gene space"}
}
\value{
A plotly visualization. Hover text contains about genes and gene clusters.
}
\description{
Plot UMAP of gene space, coloring gene clusters and marking given genes.
}
\examples{
data("pbmc_small_moduli")

pbmc_small_moduli <- run_umap_genes(pbmc_small_moduli)
visualize_gene_space(pbmc_small_moduli)

}
