% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualize_moduli_space}
\alias{visualize_moduli_space}
\title{Interactive UMAP plot of moduli space}
\usage{
visualize_moduli_space(
  moduli,
  mark.points = NULL,
  color.clusters = NULL,
  color.cluster.groups = NULL,
  title = "UMAP plot of moduli space"
)
}
\arguments{
\item{moduli}{A moduli object with a umap representation of the moduli space}

\item{mark.points}{Integer vector with ids of points to mark with "x", all points are marked with a circle if NULL.
Default value is NULL}

\item{color.clusters}{Vector with ids of analysis clusters to be colored, can only color 12 clusters. If NULL
the first 12 analysis clusters are colored. Default value is NULL}

\item{color.cluster.groups}{Vector with ids of analysis cluster groups to be colored, can only color 12 cluster groups.
Only of color.clusters or color.cluster.groups can be not NULL. Deafult value is NULL.}

\item{title}{Title of plot, default value is "UMAP plot of moduli space"}
}
\value{
A plotly visualization. Hover text contains metadata about points. If analysis clusters are
enriched, differentially expressed gene clusters are marked with an asterisk.
}
\description{
Plot UMAP of moduli space, coloring analysis clusters and marking given points.
}
\examples{
data("pbmc_small_moduli")

# clustering analysis, this is not necessary by enriches the visualization
pbmc_small_moduli <- get_snn(pbmc_small_moduli, 4)
pbmc_small_moduli <- cluster_moduli_space(pbmc_small_moduli)

# plotting
pbmc_small_moduli <- run_umap_moduli(pbmc_small_moduli)
visualize_moduli_space(pbmc_small_moduli)

}
