% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{which_point}
\alias{which_point}
\title{Retrieve ids of points}
\usage{
which_point(
  moduli,
  gene.clusters,
  filter = c("equal", "contains", "is_subset", "intersects", "in_complement")
)
}
\arguments{
\item{moduli}{A moduli object}

\item{gene.clusters}{vector with ids of gene clusters}

\item{filter}{a choice of filter, default value is \code{"equal"}. See details.}
}
\value{
A vector with point ids
}
\description{
Retrieves ids of points using a set of gene clusters.
}
\details{
The filter \code{"equal"} gets the point that represents the point representing
the clusters in \code{gene.clusters}. The option \code{"contains"} will get all points that
that representing subsets of \code{gene.clusters}, while \code{"is_subset"} will get all
points that have all elements of \code{gene.clusters}. The filter \code{"intersetcs"} retrieves
ids of points that have any gene cluster in \code{gene.clusters}, and \code{"in_complement"}
does the opposite, retrieving ids of points that only represent clusters not in \code{gene.clusters}.
}
