% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduli_clustering.R
\name{analysis_cluster_metadata}
\alias{analysis_cluster_metadata}
\title{Retrieved analysis cluster metadata}
\usage{
analysis_cluster_metadata(moduli)
}
\arguments{
\item{moduli}{A moduli object with analysis clusters save to the \code{analysis.clusters} slot}
}
\value{
A data.frame with ids, sizes, and, if available, expressed gene clusters and associated
p-values of each analysis cluster.
}
\description{
Retrieved analysis cluster metadata
}
\examples{
data("pbmc_small_moduli")
pbmc_small_moduli <- get_snn(pbmc_small_moduli, 4)
pbmc_small_moduli <- cluster_moduli_space(pbmc_small_moduli)
analysis_cluster_metadata(pbmc_small_moduli)

}
