% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduli_clustering.R
\name{cluster_moduli_space}
\alias{cluster_moduli_space}
\title{Clusters moduli space using the Leiden algorithm}
\usage{
cluster_moduli_space(
  moduli,
  resolution_parameter = 1,
  partition_type = "RBConfigurationVertexPartition",
  enrich = T,
  thld = 0.05,
  seed = 123,
  ...
)
}
\arguments{
\item{moduli}{A moduli object with a graph in the \code{snn.graph} slot}

\item{partition_type}{Type of partition to use (see \link[leiden]{leiden}), default value is
"RBConfigurationVertexPartition"}

\item{enrich}{Whether to enrich the analysis clusters with differentially expressed gene clusters
by calling \link{enrich_analyis_clusters}. Default value is TRUE}

\item{thld}{Significance threshold for gene clusters, only used if \code{enrich} is TRUE}

\item{seed}{Random seed, default value is 123}

\item{...}{Additional arguments passed to \link[leiden]{leiden}}

\item{resolution_paramenter}{Resolution passed to the Leiden algorithm (see \link[leiden]{leiden}), default
value is 1}
}
\value{
A moduli object with analysis clusters saved in the \code{analysis.clusters} slot.
}
\description{
Create clusters of analysis (i.e. points in the moduli) by applying 
\link[leiden]{leiden} to the snn graph of the moduli.
}
