% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_cluster_utils.R
\name{gene_cluster_metadata}
\alias{gene_cluster_metadata}
\title{Retrieves gene cluster metadata}
\usage{
gene_cluster_metadata(moduli)
}
\arguments{
\item{moduli}{A moduli object}
}
\value{
A data.frame
}
\description{
Retrieves gene cluster metadata
}
