% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualize_moduli_space}
\alias{visualize_moduli_space}
\title{Interactive UMAP plot of moduli space}
\usage{
visualize_moduli_space(
  moduli,
  n_neighbors = 15,
  mark.points = NULL,
  color.clusters = NULL,
  title = "UMAP plot of moduli space",
  seed = 123,
  ...
)
}
\arguments{
\item{moduli}{A moduli object}

\item{n_neighbors}{Size of local neighborhood of umap (see \link[uwot]{umap}), defaults to 15}

\item{mark.points}{Integer vector with ids of points to mark with "x", all points are marked with a circle if NULL.
Default value is NULL}

\item{color.clusters}{Vector with ids of analysis clusters to be colored, can only color 12 clusters. If NULL
the first 12 analysis clusters are colored. Default value is NULL}

\item{title}{Title of plot, dafault value is "UMAP plot of moduli space"}

\item{seed}{Random seed, default value is 123}

\item{...}{Additional paramenters passed to \link[uwot]{umap}}
}
\value{
A plotly visualization. Hover text contains metadata about points. If analysis clusters are
enriched, differentially expressed gene clusters are marked with an asterisk.
}
\description{
Plot UMAP of moduli space, coloring analysis clusters and marking given points.
}
